refactor(amounts): extract AmountInput helpers into a shared util
Move the pure, DOM-free helpers (cleanAmountInput, groupAmountInput,
caretAfterFormat, countSignificant) out of the AmountInput component into
AmountInput.util.ts, matching the feature-util convention used elsewhere
(Dashboard.utils.ts, Reconcile.utils.ts, Accounts.utils.ts). The tests now
import from the util rather than the component module.

Export groupThousands from utils/formatAmount and have groupAmountInput call
it for the integer grouping, so the comma-grouping rule lives in one place and
display and input formatting can no longer drift.
